+++ spec/unit/serializer/DreamSerializer/rendersMany.spec.ts
@@ -1,4 +1,5 @@
 import DreamSerializer from '../../../../src/serializer/DreamSerializer.js'
+import ObjectSerializer from '../../../../src/serializer/ObjectSerializer.js'
 import CalendarDate from '../../../../src/utils/datetime/CalendarDate.js'
 import Pet from '../../../../test-app/app/models/Pet.js'
 import User from '../../../../test-app/app/models/User.js'
@@ -104,7 +105,7 @@ describe('DreamSerializer#rendersMany', () => {
     })
   })
 
-  it('supports supplying a custom serializer', () => {
+  it('supports supplying a custom DreamSerializer', () => {
     const birthdate = CalendarDate.fromISO('1950-10-02')
     const user = User.new({ id: '7', name: 'Charlie', birthdate })
     const pet1 = Pet.new({ id: '3', user, name: 'Snoopy', species: 'dog' })
@@ -114,8 +115,33 @@ describe('DreamSerializer#rendersMany', () => {
     user.pets = [pet1, pet2]
 
     const CustomSerializer = (data: Pet) => DreamSerializer(Pet, data).attribute('name')
-    ;(CustomSerializer as any)['globalName'] = 'CustomPetSerializer'
-    ;(CustomSerializer as any)['openapiName'] = 'CustomPet'
+    const MySerializer = (data: User) =>
+      DreamSerializer(User, data).rendersMany('pets', { serializer: CustomSerializer })
+
+    const serializer = MySerializer(user)
+
+    expect(serializer.render()).toEqual({
+      pets: [
+        {
+          name: 'Snoopy',
+        },
+        {
+          name: 'Woodstock',
+        },
+      ],
+    })
+  })
+
+  it('supports supplying a custom ObjectSerializer', () => {
+    const birthdate = CalendarDate.fromISO('1950-10-02')
+    const user = User.new({ id: '7', name: 'Charlie', birthdate })
+    const pet1 = Pet.new({ id: '3', user, name: 'Snoopy', species: 'dog' })
+    const pet2 = Pet.new({ id: '4', user, name: 'Woodstock', species: 'frog' })
+    pet1.ratings = []
+    pet2.ratings = []
+    user.pets = [pet1, pet2]
+
+    const CustomSerializer = (data: Pet) => ObjectSerializer(data).attribute('name', { openapi: 'string' })
     const MySerializer = (data: User) =>
       DreamSerializer(User, data).rendersMany('pets', { serializer: CustomSerializer })